linting and move match sugggester to util
