+++ editor/app/ems/[incident_number]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { notFound } from "next/navigation";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
 import DataCard from "@/components/DataCard";
@@ -23,82 +23,6 @@ import EMSLinkToPersonButton, {
 import { emsMatchingPeopleColumns } from "@/configs/emsMatchingPeopleColumns";
 import { PeopleListRow } from "@/types/peopleList";
 import { FaTruckMedical } from "react-icons/fa6";
-
-interface MatchResult {
-  ems_id: number;
-  person_id: number;
-  case_id: boolean;
-  age: boolean;
-  race: boolean;
-  sex: boolean;
-  address: boolean;
-  date: boolean;
-  score: number;
-}
-
-const useMatchSuggestions = (
-  ems_pcrs?: EMSPatientCareRecord[],
-  people?: PeopleListRow[]
-): MatchResult[] =>
-  useMemo(() => {
-    const results: MatchResult[] = [];
-
-    if (!ems_pcrs || !people) {
-      return results;
-    }
-
-    ems_pcrs.forEach((ems_pcr) => {
-      people.forEach((person) => {
-        const matchResult: MatchResult = {
-          ems_id: ems_pcr.id,
-          person_id: person.id,
-          case_id: false,
-          age: false,
-          race: false,
-          sex: false,
-          address: false,
-          date: false,
-          score: 0,
-        };
-        /**
-         * Case ID
-         */
-        if (
-          person.crash?.case_id &&
-          ems_pcr.unparsed_apd_incident_numbers?.includes(person.crash.case_id)
-        ) {
-          matchResult.case_id = true;
-          matchResult.score++;
-        }
-        /**
-         * Age
-         */
-        if (
-          ems_pcr.pcr_patient_age !== null &&
-          person.prsn_age !== null &&
-          ems_pcr.pcr_patient_age === person.prsn_age
-        ) {
-          matchResult.age = true;
-          matchResult.score++;
-        }
-        /**
-         * Sex
-         */
-        if (
-          ems_pcr.pcr_patient_gender &&
-          person.gndr.label.toLowerCase() ===
-            ems_pcr.pcr_patient_gender?.toLowerCase()
-        ) {
-          matchResult.sex = true;
-          matchResult.score++;
-        }
-        results.push(matchResult);
-      });
-    });
-    return results;
-  }, [ems_pcrs, people]);
-
-const allowedLinkRecordRoles = ["vz-admin", "editor"];
 
 export default function EMSDetailsPage({
   params,
@@ -126,7 +50,7 @@ export default function EMSDetailsPage({
     typename: "ems__incidents",
   });
 
-  const { mutate: updateEMSIncident, loading: isMutating } = useMutation(
+  const { mutate: updateEMSIncident } = useMutation(
     UPDATE_EMS_INCIDENT_CRASH_AND_PERSON
   );
 
@@ -152,7 +76,6 @@ export default function EMSDetailsPage({
     typename: "people_list_view",
   });
 
-  const matchResults = useMatchSuggestions(ems_pcrs, matchingPeople);
   const onSaveCallback = useCallback(async () => {
     await refetch();
   }, [refetch]);
