+++ editor/components/EMSLinkRecordButton.tsx
@@ -14,7 +14,7 @@ export interface EMSLinkRecordButtonProps extends Record<string, unknown> {
 
 const EMSLinkRecordButton: React.FC<
   RowActionComponentProps<EMSPatientCareRecord, EMSLinkRecordButtonProps>
-> = ({ record, mutation, onSaveCallback, additionalProps }) => {
+> = ({ record, additionalProps }) => {
   //   const { mutate, loading: isMutating } = useMutation(mutation);
 
   const isLinkingInProgress = !!additionalProps?.selectedEmsPcr;
